# src/config/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"UUID": {
				"type": "string",
				"example": "52907745-7672-470e-a803-a2f8feb52944",
				"format": "uuid",
				"description": "Stringified UUIDv4.\r\nSee [RFC 4112](https://tools.ietf.org/html/rfc4122)",
				"pattern": "[0-9A-Fa-f]{8}-[0-9A-Fa-f]{4}-4[0-9A-Fa-f]{3}-[89ABab][0-9A-Fa-f]{3}-[0-9A-Fa-f]{12}"
			},
...
			"ErrorDetail": {
				"properties": {
					"field": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"value": {
						"type": "string"
					}
				},
				"required": [
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ForbiddenError": {
				"properties": {
					"name": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"stack": {
						"type": "string"
					},
					"status": {
						"type": "number",
						"format": "double"
					},
					"code": {
						"type": "number",
						"format": "double"
					},
					"details": {
						"items": {
							"$ref": "#/components/schemas/ErrorDetail"
						},
						"type": "array"
					}
				},
				"required": [
					"name",
					"message",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UnauthorizedError": {
				"properties": {
					"name": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"stack": {
						"type": "string"
					},
					"status": {
						"type": "number",
						"format": "double"
					},
					"code": {
						"type": "number",
						"format": "double"
					},
					"details": {
						"items": {
							"$ref": "#/components/schemas/ErrorDetail"
						},
						"type": "array"
					}
				},
				"required": [
					"name",
					"message",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"DefaultString": {
				"type": "string",
				"example": "string",
				"minLength": 2,
				"maxLength": 50
			},
...
			"BacklogCompleteOut": {
				"properties": {
					"id": {
						"$ref": "#/components/schemas/UUID"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"deletedAt": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"branchId": {
						"allOf": [
							{
								"$ref": "#/components/schemas/UUID"
							}
						],
						"nullable": true
					},
					"categories": {
						"items": {
							"$ref": "#/components/schemas/CategoryCompleteOut"
						},
						"type": "array",
						"nullable": true
					}
				},
				"required": [
					"id",
					"createdAt",
					"updatedAt",
					"branchId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"BranchNotFound": {
				"properties": {
					"name": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"stack": {
						"type": "string"
					},
					"status": {
						"type": "number",
						"format": "double"
					},
					"code": {
						"type": "number",
						"format": "double"
					},
					"details": {
						"items": {
							"$ref": "#/components/schemas/ErrorDetail"
						},
						"type": "array"
					}
				},
				"required": [
					"name",
					"message",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateMenuCompleteOut": {
				"properties": {
					"id": {
						"$ref": "#/components/schemas/UUID"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"deletedAt": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"branchId": {
						"allOf": [
							{
								"$ref": "#/components/schemas/UUID"
							}
						],
						"nullable": true
					},
					"name": {
						"allOf": [
							{
								"$ref": "#/components/schemas/DefaultString"
							}
						],
						"nullable": true
					},
					"favicon": {
						"allOf": [
							{
								"$ref": "#/components/schemas/DefaultString"
							}
						],
						"nullable": true
					},
					"isPublished": {
						"allOf": [
							{
								"$ref": "#/components/schemas/Boolean"
							}
						],
						"nullable": true
					},
					"restaurantId": {
						"allOf": [
							{
								"$ref": "#/components/schemas/DefaultString"
							}
						],
						"nullable": true
					}
				},
				"required": [
					"id",
					"createdAt",
					"updatedAt",
					"branchId",
					"name",
					"favicon",
					"isPublished"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"SessionAuth": {
				"type": "apiKey",
				"in": "cookie",
				"name": "session-id"
			},
			"JWTAuth": {
				"type": "apiKey",
				"in": "cookie",
				"name": "access-token"
			}
		}
	},
	"info": {
		"title": "Menuchi-API",
		"version": "1.0.0",
		"description": "Menuchi API Documentation",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
...
		"/menus/backlog/{backlogId}": {
			"get": {
				"operationId": "GetBacklog",
				"responses": {
					"200": {
						"description": "Backlog is retrieved successfully.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BacklogCompleteOut"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized user.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UnauthorizedError"
								}
							}
						}
					},
					"403": {
						"description": "Access Denied. You are not authorized to perform this action.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ForbiddenError"
								}
							}
						}
					},
					"404": {
						"description": "4049 BranchNotFound",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BranchNotFound"
								}
							}
						}
					}
				},
				"description": "Retrieves a backlog by ID.",
				"tags": [
					"Menu"
				],
				"security": [
					{
						"": [
							"RESTAURANT_OWNER"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "backlogId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/UUID"
						}
					},
					{
						"in": "query",
						"name": "search",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/DefaultString"
						}
					}
				]
			}
		},
		"/menus": {
			"post": {
				"operationId": "CreateMenu",
				"responses": {
					"201": {
						"description": "Menu created successfully.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CreateMenuCompleteOut"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized user.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UnauthorizedError"
								}
							}
						}
					},
					"403": {
						"description": "Access Denied. You are not authorized to perform this action.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ForbiddenError"
								}
							}
						}
					},
					"404": {
						"description": "4049 BranchNotFound",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BranchNotFound"
								}
							}
						}
					}
				},
				"description": "Creates a new menu.",
				"tags": [
					"Menu"
				],
				"security": [
					{
						"": [
							"RESTAURANT_OWNER"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateMenuCompactIn"
							}
						}
					}
				}
			}
		},
		"/menus/{menuId}": {
			"get": {
				"operationId": "GetMenu",
				"responses": {
					"200": {
						"description": "Menu is retrieved successfully.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MenuCompleteOut"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized user.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UnauthorizedError"
								}
							}
						}
					},
					"403": {
						"description": "Access Denied. You are not authorized to perform this action.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ForbiddenError"
								}
							}
						}
					},
					"404": {
						"description": "4045 MenuNotFound",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MenuNotFound"
								}
							}
						}
					}
				},
				"description": "Retrieves a menu by ID.",
				"tags": [
					"Menu"
				],
				"security": [
					{
						"": [
							"RESTAURANT_OWNER"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "menuId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/UUID"
						}
					}
				]
			},
			"patch": {
				"operationId": "UpdateMenu",
				"responses": {
					"204": {
						"description": "Menu updated successfully.",
						"content": {
							"application/json": {
								"schema": {
									"type": "number",
									"enum": [
										null
									],
									"nullable": true
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized user.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UnauthorizedError"
								}
							}
						}
					},
					"403": {
						"description": "Access Denied. You are not authorized to perform this action.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ForbiddenError"
								}
							}
						}
					},
					"404": {
						"description": "4045 MenuNotFound",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MenuNotFound"
								}
							}
						}
					},
					"422": {
						"description": "4225 MenuValidationError",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MenuValidationError"
								}
							}
						}
					}
				},
				"description": "Updates a menu.",
				"tags": [
					"Menu"
				],
				"security": [
					{
						"": [
							"RESTAURANT_OWNER"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "menuId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/UUID"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/MenuCompactIn"
							}
						}
					}
				}
			}
		},
		"/menus/{menuId}/cylinders": {
			"post": {
				"operationId": "CreateCylinder",
				"responses": {
					"201": {
						"description": "Cylinder created successfully.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CreateCylinderCompleteOut"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized user.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UnauthorizedError"
								}
							}
						}
					},
					"403": {
						"description": "Access Denied. You are not authorized to perform this action.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ForbiddenError"
								}
							}
						}
					},
					"404": {
						"description": "4045 MenuNotFound",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MenuNotFound"
								}
							}
						}
					},
					"409": {
						"description": "ConstraintsDatabaseError",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ConstraintsDatabaseError"
								}
							}
						}
					},
					"422": {
						"description": "4226 CylinderValidationError",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CylinderValidationError"
								}
							}
						}
					}
				},
				"description": "Creates a new cylinder in a menu.",
				"tags": [
					"Menu"
				],
				"security": [
					{
						"": [
							"RESTAURANT_OWNER"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "menuId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/UUID"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CylinderCompactIn"
							}
						}
					}
				}
			},
			"patch": {
				"operationId": "ReorderCylinders",
				"responses": {
					"204": {
						"description": "Cylinder orders in the menu updated successfully.",
						"content": {
							"application/json": {
								"schema": {
									"type": "number",
									"format": "double"
								}
							}
						}
					},
					"400": {
						"description": "All menu cylinder IDs must be in the request.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MenuchiError"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized user.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UnauthorizedError"
								}
							}
						}
					},
					"403": {
						"description": "Access Denied. You are not authorized to perform this action.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ForbiddenError"
								}
							}
						}
					}
				},
				"description": "Reorders cylinders.",
				"tags": [
					"Menu"
				],
				"security": [
					{
						"": [
							"RESTAURANT_OWNER"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "menuId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/UUID"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"items": {
									"$ref": "#/components/schemas/UUID"
								},
								"type": "array"
							}
						}
					}
				}
			}
		},
		"/menus/{menuId}/categories": {
			"post": {
				"operationId": "CreateMenuCategory",
				"responses": {
					"201": {
						"description": "Menu Category created successfully.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CreateMenuCategoryCompleteOut"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized user.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UnauthorizedError"
								}
							}
						}
					},
					"403": {
						"description": "Access Denied. You are not authorized to perform this action.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ForbiddenError"
								}
							}
						}
					},
					"409": {
						"description": "ConstraintsDatabaseError",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ConstraintsDatabaseError"
								}
							}
						}
					},
					"422": {
						"description": "4227 MenuCategoryValidationError",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MenuCategoryValidationError"
								}
							}
						}
					}
				},
				"description": "Creates a new category in a menu.",
				"tags": [
					"Menu"
				],
				"security": [
					{
						"": [
							"RESTAURANT_OWNER"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "menuId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/UUID"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/MenuCategoryCompactIn"
							}
						}
					}
				}
			},
			"patch": {
				"operationId": "ReorderMenuItems",
				"responses": {
					"204": {
						"description": "Menu item orders in the menu category updated successfully.",
						"content": {
							"application/json": {
								"schema": {
									"type": "number",
									"format": "double"
								}
							}
						}
					},
					"400": {
						"description": "All menu item IDs must be in the request.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MenuchiError"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized user.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UnauthorizedError"
								}
							}
						}
					},
					"403": {
						"description": "Access Denied. You are not authorized to perform this action.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ForbiddenError"
								}
							}
						}
					}
				},
				"description": "Reorders menu items.",
				"tags": [
					"Menu"
				],
				"security": [
					{
						"": [
							"RESTAURANT_OWNER"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "menuId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/UUID"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"items": {
									"$ref": "#/components/schemas/UUID"
								},
								"type": "array"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteMenuCategory",
				"responses": {
					"204": {
						"description": "Menu Categories deleted successfully.",
						"content": {
							"application/json": {
								"schema": {
									"type": "number",
									"enum": [
										null
									],
									"nullable": true
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized user.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UnauthorizedError"
								}
							}
						}
					},
					"403": {
						"description": "Access Denied. You are not authorized to perform this action.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ForbiddenError"
								}
							}
						}
					}
				},
				"description": "Deletes menu categories.",
				"tags": [
					"Menu"
				],
				"security": [
					{
						"": [
							"RESTAURANT_OWNER"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "menuId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/UUID"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"items": {
									"$ref": "#/components/schemas/UUID"
								},
								"type": "array"
							}
						}
					}
				}
			}
		},
		"/menus/{menuId}/items": {
			"patch": {
				"operationId": "ReorderMenuCategories",
				"responses": {
					"204": {
						"description": "Menu category orders in the cylinder updated successfully.",
						"content": {
							"application/json": {
								"schema": {
									"type": "number",
									"format": "double"
								}
							}
						}
					},
					"400": {
						"description": "All menu category IDs must be in the request.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MenuchiError"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized user.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UnauthorizedError"
								}
							}
						}
					},
					"403": {
						"description": "Access Denied. You are not authorized to perform this action.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ForbiddenError"
								}
							}
						}
					}
				},
				"description": "Reorders menu categories.",
				"tags": [
					"Menu"
				],
				"security": [
					{
						"": [
							"RESTAURANT_OWNER"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "menuId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/UUID"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"items": {
									"$ref": "#/components/schemas/UUID"
								},
								"type": "array"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteMenuItem",
				"responses": {
					"204": {
						"description": "Menu Items deleted successfully.",
						"content": {
							"application/json": {
								"schema": {
									"type": "number",
									"enum": [
										null
									],
									"nullable": true
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized user.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UnauthorizedError"
								}
							}
						}
					},
					"403": {
						"description": "Access Denied. You are not authorized to perform this action.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ForbiddenError"
								}
							}
						}
					}
				},
				"description": "Deletes menu items.",
				"tags": [
					"Menu"
				],
				"security": [
					{
						"": [
							"RESTAURANT_OWNER"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "menuId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/UUID"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"items": {
									"$ref": "#/components/schemas/UUID"
								},
								"type": "array"
							}
						}
					}
				}
			}
		},
		"/menus/{menuId}/items/{menuItemId}/hide/{isHide}": {
			"patch": {
				"operationId": "HideMenuItem",
				"responses": {
					"204": {
						"description": "Menu Item hide/unhide successfully.",
						"content": {
							"application/json": {
								"schema": {
									"type": "number",
									"enum": [
										null
									],
									"nullable": true
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized user.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UnauthorizedError"
								}
							}
						}
					},
					"403": {
						"description": "Access Denied. You are not authorized to perform this action.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ForbiddenError"
								}
							}
						}
					}
				},
				"description": "Hide/Unhide a menu item.",
				"tags": [
					"Menu"
				],
				"security": [
					{
						"": [
							"RESTAURANT_OWNER"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "menuId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/UUID"
						}
					},
					{
						"in": "path",
						"name": "menuItemId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/UUID"
						}
					},
					{
						"in": "path",
						"name": "isHide",
						"required": true,
						"schema": {
							"type": "boolean"
						}
					}
				]
			}
		},
...
		"/backlog/{backlogId}": {
			"get": {
				"operationId": "GeBacklog",
				"responses": {
					"200": {
						"description": "Backlog is retrieved successfully.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BacklogCompleteOut"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized user.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UnauthorizedError"
								}
							}
						}
					},
					"403": {
						"description": "Access Denied. You are not authorized to perform this action.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ForbiddenError"
								}
							}
						}
					},
					"404": {
						"description": "4044 BacklogNotFound",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BacklogNotFound"
								}
							}
						}
					}
				},
				"description": "Retrieves a specific backlog.",
				"tags": [
					"Backlog"
				],
				"security": [
					{
						"": [
							"RESTAURANT_OWNER"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "backlogId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/UUID"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	],
...
}